@@ -1,12 +1,15 @@
 package com.nitramite.adapters;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.nitramite.nitspec.R;
+
 import java.util.List;
 
 public class DeviceItemsAdapter extends RecyclerView.Adapter<DeviceItemsAdapter.MyViewHolder> {
